requirements.txt
asgiref==3.3.1
Django==3.0.5
django-birthday==0.1.3
django-cors-headers==3.7.0
django-qr-code==2.1.0
djangorestframework==3.12.2
djongo==1.3.3
pymongo==3.11.2
pytz==2020.5
segno==1.3.1
sqlparse==0.2.4
